global: SSS22.1, Clang stability (works), misc features
Mess again... I really have to review my workflow.

Post-Clang fixes (now it works stable!):
* temporary build xrGame with precise math until I move physics-related
  files to a standalone folder/subproject (fastmath prohibits usage of
  NaN and infinity, which physics code uses a lot);
* removed ugly hack with g_graph_engine that didn't work anyway --
  problem_solver_inline_ext.h now allows including ai_space.h without
  include hell even with Clang.

Graphics:
* brand-new shader cache, fast and reliable. We always run a preprocess
  pass and then calculate xxHash for it. If it's the same as in the
  cache, this means that the compiled shader will be 100% the same. If
  you had any changes in either the shader itself or in one of the
  header files it includes, the preprocessed data will differ, so that
  the shader will be recompiled, but only this one.
  This is way more reliable than GSC cache (Anomaly has the same) and
  faster than OXR (as they hash the whole shaders/ directory, which
  means any change in the folder will trigger a recompilation of all
  the shaders, even if you only removed a comment or an empty line;

* OGSR-specific things like handling HUD mask drops, visor reflections
  etc., was removed. I'll stick to the Anomaly implementations for now,
  as my first priority is to make SSS look as they're meant to look.
  Same for engine logics for calculating rain drops, wet surfaces, and
  dynamic DoF -- I'd like to tune them from the scripts (still prefer
  Lua much more than C++).

* the latest SSS22.1;
* Z terrain prepass and geometry optimization from Anomaly;
* full complex color grading (8 4D parameters for ACES tonemapping);
* shadows filtering and mblur from Anomaly Shaders Look Better;
* ability to turn on both volumetric and flat sunshafts at the same time
  ("combined", like in Anom). Volumetic + Manowar look really eye-candy;
* remove a ton of legacy code: non-DX11 paths, workarounds for cards
  from 2005, never active/used/usable options etc. etc.;
* use DirectX 11.2 API to create HW contexts, chain descs etc. If/when
  upstream OGSR proposes a better solution, I'll pick it.
* use oneTBB harder: HOMs, texture loading (incl. thm parsing), skinning,
  particle rendering, etc.;
* actor shadow from the latest OXR.

Some stuff like sunshafts_min, gloss_factor_min etc., as well as texture
material parsing picked from Anomaly to match it as close as possible.
As the initial idea was to be able to use the latest Atmospherics with
on SoC mod :>

My main plan is to keep the engine up to date with the latest SSS as
soon as they get releaed, BUT using the most advanced (from my PoV) OXR
code to render as much stuff in parallel as possible (sun cascades and
rain are already implemented, so I have a good example. I'm also
thinking of details/grass and probably more SSS-specific effects, as
most of them has their own RTs).

Sound:
* float32 sound support from OXR (sounds really great!);
* Doppler/smoothing from Anomaly Modded Exes;
* improved sound occlusion from OpenXRay (accounting materials);
* fixed sound environment zones broken in OGSR for now reason.

Waiting for the SteamAudio integration in OXR :>
TBH I'd love to switch from OGSR to OXR in the future, when the latter
one gets full stable SoC support.

Weather/environment:
* ability to handle CoP's sky stretch from the configs. Let's say the
  original Atmosfear skies still used a lot in most mods were rendered
  to account this stretching, but in most engines the stretch was fixed.
  IOW Atmosfear skies looks "squashed" there and nobody cares (except
  me).
* now we don't load the entire textures/sky/ when the game starts. I
  have 3.5 Gb+ textures there and it makes no sense to keep them as long
  as the game runs as you'll barely see 20% of them in one game day.

Misc:
* introduce and use a couple memcpy helpers which use 128 and 256 bit
  AVX registers to copy data and use it when updating shader constants.
* align a couple more classes to 16 bytes, use custom constructors for
  them to always use 128-bit assignments;
* lots of optimizations from OpenXRay ^.^

System reqs are the following, but I really hope most people working
with such advanced engines already has all this:

* Windows 10+ (it may work on 8.1, but no guarantees)
* AVX2-compatible processor (which means a 2013 year old one or any
  newer)
* DirectX 11.2 (Win10 has it anyway?)

Note: cflags are currently hardcoded for -malderlake -mavxvnni. I'll
remove that once I switch the build system to CMake, where the CPU
selection will be available via standard options.

+++ ogsr_engine/Layers/xrRenderDX10/dx10r_constants_cache.h
@@ -98,7 +98,7 @@ public:
         flush_cache();
     }
 
-    ICF void access_direct(R_constant* C, u32 DataSize, void** ppVData, void** ppGData, void** ppPData)
+    ICF void access_direct(R_constant* C, size_t DataSize, void** ppVData, void** ppGData, void** ppPData)
     {
         if (ppPData)
         {
